Pointers in C
Pointers store address of variables or a memory location.

To access address of a variable to a pointer, we use the unary operator & (ampersand) that returns the address of that variable. For example &x gives us address of variable x.

One more operator is unary * (Asterisk) which is used for two things :
To declare a pointer variable: When a pointer variable is declared in C/C++, there must a * before its name.

To access the value stored in the address we use the unary operator (*) that returns the value of the variable located at the address specified by its operand.